requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
scikit-learn
streamlit
requests